requirements.txt
jupyter_core>=4.4.0
jupyter_client>=5.2.0
notebook>=5.3.0,<6.0
traitlets>=4.2.0
tornado>=4.2.0
requests>=2.7,<3.0
redis==3.1.0
PyYAML==3.13

